change schema plugin to be returned from function (#3252)
* change schema plugin to be returned from function

* fix forced-layout example
