+++ packages/slate-schema/test/index.js
@@ -1,12 +1,13 @@
 import assert from 'assert'
 import { fixtures } from '../../../support/fixtures'
 import { Editor } from 'slate'
-import { withSchema } from '..'
+import { defineSchema } from '..'
 
 describe('slate-schema', () => {
   fixtures(__dirname, 'validations', ({ module }) => {
     const { input, schema, output } = module
-    const editor = withSchema(input, schema)
+    const withSchema = defineSchema(schema)
+    const editor = withSchema(input)
     Editor.normalize(editor, { force: true })
     assert.deepEqual(editor.children, output.children)
   })
